RecordExpressionParser: splitted into several files
