+++ core/record/recordExpressionParser/index.js
@@ -1,0 +1,1 @@
+export { evalApplicableExpression, evalApplicableExpressions, evalNodeQuery } from './recordExpressionParser'
