+++ core/record/recordExpressionParser/recordEpressionFunctions.js
